+++ CHANGELOG.md
@@ -86,6 +86,21 @@ Reconstruction sampler:
   "Review interval" button.
 
 
+Layouts:
+
+- The Stack View section of the Settings Widget allows now the configuration of
+  a list of default layouts that can be applied to newly opened stacks and stack
+  groups. At the moment this is mainly useful to organize orthogonal views in a
+  custom way, but will be extended to support custom project layouts.
+
+- By default only one layout is available: organize XY, XZ and ZY in four
+  quadrants using one additional window (if non available the help page). This
+  is its specification: h(v(XY, XZ), v(ZY, o(F1))). With h() and v(), horizontal
+  and vertical splits are declared, respectively. With o(F1) a help window will
+  be opened as fourth window if not already another window exists..
+
+
+
 Miscellaneous:
 
 - If multiple tile layers are used, the stack viewer layer settings (blue white